+++ src/collective/fhirpath/utils.py
@@ -4,12 +4,13 @@ from fhirpath.storage import MemoryStorage
 from plone import api
 from plone.app.fhirfield.interfaces import IFhirResource
 from plone.behavior.interfaces import IBehavior
+from plone.restapi.services import _no_content_marker
+from pydantic import BaseModel
 from zope.component import getUtility
 from zope.schema import getFields
-from plone.restapi.services import _no_content_marker
+
 import json
 import os
-from pydantic import BaseModel
 
 
 __author__ = "Md Nazrul Islam<email2nazrul@gmail.com>"
@@ -119,9 +120,7 @@ class FHIRModelServiceMixin:
     def render(self):
         """ """
         self.check_permission()
-
         content = self.reply()
-
         if content is _no_content_marker:
             return